refactor : category page
파일명, 폴더명, api, style refactoring

Related to : #149
