+++ src/styles/category/category.module.css
@@ -1,15 +1,5 @@
+/* cart page wrapper */
 .container {
-	/* width: 100vw; */
-
-	/* display: flex;
-	justify-content: center; */
-	display: grid;
-	grid-template-columns: 1fr 1300px 1fr;
-}
-.wrapper {
-	width: 1300px;
-
 	display: flex;
-	flex-direction: row;
-	justify-content: space-between;
+	flex-direction: column;
 }
