--- src/styles/category/categoryContents/categorySlider/categorySlider.module.css
@@ -1,5 +1,0 @@
-.categorySliderWrapper {
-	width: 1025px;
-	height: 250px;
-	margin-bottom: 50px;
-}
