--- src/styles/category/categoryContents/categorySlider/customArrow.module.css
@@ -1,59 +1,0 @@
-.arrow {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	position: absolute;
-	/* width: 40px; */
-	/* height: 40px; */
-	border-radius: 50%;
-	/* box-sizing: border-box; */
-	background-color: #615a5f;
-	z-index: 10;
-	/* transition:
-		width 0.2s ease,
-		height 0.2s ease,
-		left 0.2s ease,
-		right 0.2s ease; */
-}
-
-.arrowPrev {
-	left: 10px;
-}
-
-.arrowNext {
-	right: 10px;
-}
-
-.arrow:hover {
-	background-color: #454144;
-}
-
-.arrowPrev:hover {
-	left: 5px; /* 원래 스타일로 되돌립니다. */
-}
-
-.arrowNext:hover {
-	right: 5px; /* 원래 스타일로 되돌립니다. */
-}
-
-/* 
-.arrow:hover {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	position: absolute;
-	width: 47px;
-	height: 47px;
-	border-radius: 50%;
-	box-sizing: border-box;
-	background-color: #f4f2e5;
-	z-index: 1;
-}
-
-.arrowPrev:hover {
-	left: calc(180px - ((45px - 38px) / 2));
-}
-
-.arrowNext:hover {
-	right: calc(180px - ((45px - 38px) / 2));
-}  */
