--- src/styles/category/categoryContents/categorySlider/sliderItem.module.css
@@ -1,27 +1,0 @@
-.sliderItem {
-	width: 1025px;
-	height: 250px;
-
-	display: flex;
-}
-.imageBox {
-	width: 600px;
-	height: 250px;
-
-	display: flex;
-	justify-content: space-around;
-	align-items: center;
-
-	overflow: hidden;
-
-	background-color: #90898e;
-	background-color: #eae8fa;
-}
-.bookImage {
-	box-shadow: 10px 10px 5px 0px rgba(0, 0, 0, 0.35);
-	transform: rotate(-25deg) translateY(10px);
-}
-.textBox {
-	width: 425px;
-	height: 250px;
-}
