+++ src/styles/category/categoryView/categoryContents/categoryContents.module.css
@@ -1,0 +1,10 @@
+.categoryContents {
+	width: 1025px;
+	margin-top: 50px;
+	margin-bottom: 50px;
+}
+
+/* .categoryContents > div:last-child {
+	margin-bottom: 60px;
+	background-color: red;
+} */
