+++ src/styles/category/categoryView/categoryContents/categorySlider/categorySlider.module.css
@@ -1,0 +1,5 @@
+.categorySliderWrapper {
+	width: 1025px;
+	height: 250px;
+	margin-bottom: 50px;
+}
