+++ src/styles/category/categoryView/categoryContents/sortBar/sortBar.module.css
@@ -1,0 +1,44 @@
+.sortBox {
+	width: inherit;
+	height: 60px;
+
+	border: solid 1px #ebebeb;
+
+	display: flex;
+	justify-content: space-between;
+}
+.textBox {
+	/* width: 250px; */
+	height: inherit;
+
+	display: flex;
+	/* flex-direction: row-reverse; */
+	align-items: center;
+}
+.textBox > p {
+	font-size: 12px;
+	margin-right: 25px;
+
+	cursor: pointer;
+
+	color: #666666;
+	/* color: #999; */
+}
+.textBox > p:hover {
+	text-decoration: underline;
+}
+
+/* 선택된 sort item */
+.selectedSortItem {
+	font-weight: bold;
+}
+
+.amountBox {
+	width: 100px;
+	height: 60px;
+	font-size: 14px;
+
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
